backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
supabase==2.0.2
groq==0.4.2
openai==1.3.7
langchain==0.0.350
langchain-groq==0.0.1
chromadb==0.4.18
sentence-transformers==2.2.2
pypdf2==3.0.1
python-docx==1.1.0
redis==5.0.1
celery==5.3.4
httpx==0.25.2
python-magic==0.4.27
Pillow==10.1.0
